+++ CHANGELOG.md
@@ -7,7 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 
-# [0.4.11] - 2022-12-15
+## [0.4.12] - 2023-01-31
+### Added
+- Added `view_type_time_sleep` to the Genesys `queue_performance_detail_view`.
+
+### Changed
+- Updated `genesys_to_adls.py` flow with the `adls_bulk_upload` task
+- Updated `mindful_to_adls.py` flow with the `adls_bulk_upload` task
+- Changed `MindfulToCSV` task to download surveys info.
+
+
+## [0.4.11] - 2022-12-15
 ### Added
 - Added into `Genesys` the new view type `AGENT`.